This commit was generated by cvs2svn to compensate for changes in r8363, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@8364 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/drive/cia1581d.c
@@ -180,7 +180,7 @@ static inline void store_ciapb(drive_context_t *ctxptr, CLOCK rclk, BYTE byte)
                 | (iec_info->cpu_port >> 7)
                 | ((iec_info->cpu_bus << 3) & 0x80));
         } else {
-            ctxptr->func.iec_write(~byte);
+            ctxptr->func.iec_write((BYTE)(~byte));
         }
     }
 }
